+++ NEWS.md
@@ -3,6 +3,8 @@
 * Method `Duplicates$detectDuplicates()` now has arguments `n` (passed into 
 polmineR::ngrams) and `character_selection`. Values were hard-coded previously.
 Default values are aligned with Kliche et al. 2014.
+* Remaining use of `CQI$struc2cpos()` has been replaced by RcppCWB functionality
+(`get_region_matrix()`) #5.
 
 ## v0.1.9